feat: JWT 토큰 및 사용자 정보 모델을 schemas.py로 이동
